+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmImpl.kt
@@ -244,7 +244,7 @@ public class RealmImpl private constructor(
         }
     }
 
-    override fun <T, C> registerObserver(t: Thawable<Observable<T, C>>): Flow<C> {
+    override fun <T : Observable<T, C>, C> registerObserver(t: NotificationFlowable<T, C>): Flow<C> {
         return notifier.registerObserver(t)
     }