+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmSetInternal.kt
@@ -17,6 +17,7 @@
 package io.realm.kotlin.internal
 
 import io.realm.kotlin.UpdatePolicy
+import io.realm.kotlin.Versioned
 import io.realm.kotlin.internal.RealmValueArgumentConverter.convertToQueryArgs
 import io.realm.kotlin.internal.interop.Callback
 import io.realm.kotlin.internal.interop.ClassKey
@@ -38,8 +39,7 @@ import io.realm.kotlin.notifications.internal.UpdatedSetImpl
 import io.realm.kotlin.query.RealmQuery
 import io.realm.kotlin.types.BaseRealmObject
 import io.realm.kotlin.types.RealmSet
-import kotlinx.coroutines.channels.ChannelResult
-import kotlinx.coroutines.channels.SendChannel
+import kotlinx.coroutines.channels.ProducerScope
 import kotlinx.coroutines.flow.Flow
 import kotlin.reflect.KClass
 
@@ -63,7 +63,7 @@ internal class ManagedRealmSet<E> constructor(
     internal val parent: RealmObjectReference<*>,
     internal val nativePointer: RealmSetPointer,
     val operator: SetOperator<E>
-) : AbstractMutableSet<E>(), RealmSet<E>, InternalDeleteable, Observable<ManagedRealmSet<E>, SetChange<E>> {
+) : AbstractMutableSet<E>(), RealmSet<E>, InternalDeleteable, Observable<ManagedRealmSet<E>, SetChange<E>>, Versioned by operator.realmReference {
 
     override val size: Int
         get() {
@@ -129,7 +129,6 @@ internal class ManagedRealmSet<E> constructor(
     }
 
     override fun asFlow(): Flow<SetChange<E>> {
-        operator.realmReference.checkClosed()
         return operator.realmReference.owner.registerObserver(this)
     }
 
@@ -151,27 +150,49 @@ internal class ManagedRealmSet<E> constructor(
         return RealmInterop.realm_set_add_notification_callback(nativePointer, callback)
     }
 
-    override fun emitFrozenUpdate(
-        frozenRealm: RealmReference,
-        change: RealmChangesPointer,
-        channel: SendChannel<SetChange<E>>
-    ): ChannelResult<Unit>? {
-        val frozenSet: ManagedRealmSet<E>? = freeze(frozenRealm)
-        return if (frozenSet != null) {
+    override fun observable(
+        liveRealm: LiveRealm,
+        channel: ProducerScope<SetChange<E>>
+    ): NotificationFlow<ManagedRealmSet<E>, SetChange<E>> = object : NotificationFlow<ManagedRealmSet<E>, SetChange<E>>(liveRealm, this, channel) {
+        override fun initial(frozenRef: ManagedRealmSet<E>): SetChange<E> = InitialSetImpl(frozenRef)
+
+        override fun update(
+            frozenRef: ManagedRealmSet<E>,
+            change: RealmChangesPointer
+        ): SetChange<E>? {
             val builder = SetChangeSetBuilderImpl(change)
 
-            if (builder.isEmpty()) {
-                channel.trySend(InitialSetImpl(frozenSet))
+            return if (!builder.isEmpty()) {
+                UpdatedSetImpl(frozenRef, builder.build())
             } else {
-                channel.trySend(UpdatedSetImpl(frozenSet, builder.build()))
+                null
             }
-        } else {
-            channel.trySend(DeletedSetImpl(UnmanagedRealmSet()))
-                .also {
-                    channel.close()
-                }
         }
+
+        override fun delete(): SetChange<E> = DeletedSetImpl(UnmanagedRealmSet())
     }
+
+//    override fun changeEvent(
+//        frozenRealm: RealmReference,
+//        change: RealmChangesPointer,
+//        channel: SendChannel<SetChange<E>>
+//    ): ChannelResult<Unit>? {
+//        val frozenSet: ManagedRealmSet<E>? = freeze(frozenRealm)
+//        return if (frozenSet != null) {
+//            val builder = SetChangeSetBuilderImpl(change)
+//
+//            if (builder.isEmpty()) {
+//                channel.trySend(InitialSetImpl(frozenSet))
+//            } else {
+//                channel.trySend(UpdatedSetImpl(frozenSet, builder.build()))
+//            }
+//        } else {
+//            channel.trySend(DeletedSetImpl(UnmanagedRealmSet()))
+//                .also {
+//                    channel.close()
+//                }
+//        }
+//    }
 
     override fun delete() {
         RealmInterop.realm_set_remove_all(nativePointer)