+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/query/ObjectQuery.kt
@@ -17,14 +17,14 @@
 package io.realm.kotlin.internal.query
 
 import io.realm.kotlin.internal.CoreExceptionConverter
-import io.realm.kotlin.internal.Flowable
 import io.realm.kotlin.internal.InternalDeleteable
+import io.realm.kotlin.internal.LiveRealm
 import io.realm.kotlin.internal.Mediator
-import io.realm.kotlin.internal.Observable
+import io.realm.kotlin.internal.NotificationFlow
+import io.realm.kotlin.internal.NotificationFlowable
 import io.realm.kotlin.internal.RealmReference
 import io.realm.kotlin.internal.RealmResultsImpl
 import io.realm.kotlin.internal.RealmValueArgumentConverter.convertToQueryArgs
-import io.realm.kotlin.internal.Thawable
 import io.realm.kotlin.internal.asInternalDeleteable
 import io.realm.kotlin.internal.interop.ClassKey
 import io.realm.kotlin.internal.interop.RealmCoreException
@@ -44,6 +44,7 @@ import io.realm.kotlin.query.RealmScalarQuery
 import io.realm.kotlin.query.RealmSingleQuery
 import io.realm.kotlin.query.Sort
 import io.realm.kotlin.types.BaseRealmObject
+import kotlinx.coroutines.channels.ProducerScope
 import kotlinx.coroutines.flow.Flow
 import kotlin.reflect.KClass
 
@@ -54,7 +55,7 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
     private val clazz: KClass<E>,
     private val mediator: Mediator,
     internal val queryPointer: RealmQueryPointer,
-) : RealmQuery<E>, InternalDeleteable, Thawable<Observable<RealmResultsImpl<E>, ResultsChange<E>>>, Flowable<ResultsChange<E>> {
+) : RealmQuery<E>, InternalDeleteable, NotificationFlowable<RealmResultsImpl<E>, ResultsChange<E>> {
 
     private val resultsPointer: RealmResultsPointer by lazy {
         RealmInterop.realm_query_find_all(queryPointer)
@@ -176,11 +177,13 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
     override fun count(): RealmScalarQuery<Long> =
         CountQuery(realmReference, queryPointer, mediator, classKey, clazz)
 
-    override fun thaw(liveRealm: RealmReference): RealmResultsImpl<E> =
-        thawResults(liveRealm, resultsPointer, classKey, clazz, mediator)
+    override fun observable(
+        liveRealm: LiveRealm,
+        channel: ProducerScope<ResultsChange<E>>
+    ): NotificationFlow<RealmResultsImpl<E>, ResultsChange<E>> =
+        thawResults(liveRealm.realmReference, resultsPointer, classKey, clazz, mediator).observable(liveRealm, channel)
 
     override fun asFlow(): Flow<ResultsChange<E>> {
-        realmReference.checkClosed()
         return realmReference.owner
             .registerObserver(this)
     }