+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/query/SingleQuery.kt
@@ -2,30 +2,32 @@ package io.realm.kotlin.internal.query
 
 import io.realm.kotlin.ext.asFlow
 import io.realm.kotlin.internal.InternalDeleteable
+import io.realm.kotlin.internal.LiveRealm
 import io.realm.kotlin.internal.Mediator
-import io.realm.kotlin.internal.Observable
+import io.realm.kotlin.internal.NotificationFlow
+import io.realm.kotlin.internal.NotificationFlowable
 import io.realm.kotlin.internal.RealmReference
 import io.realm.kotlin.internal.RealmResultsImpl
-import io.realm.kotlin.internal.Thawable
-import io.realm.kotlin.internal.hasSameObjectKey
 import io.realm.kotlin.internal.interop.ClassKey
 import io.realm.kotlin.internal.interop.Link
 import io.realm.kotlin.internal.interop.RealmInterop
 import io.realm.kotlin.internal.interop.RealmQueryPointer
+import io.realm.kotlin.internal.realmObjectReference
 import io.realm.kotlin.internal.runIfManaged
 import io.realm.kotlin.internal.toRealmObject
-import io.realm.kotlin.notifications.InitialResults
 import io.realm.kotlin.notifications.ResultsChange
 import io.realm.kotlin.notifications.SingleQueryChange
-import io.realm.kotlin.notifications.UpdatedResults
-import io.realm.kotlin.notifications.internal.Cancellable
+import io.realm.kotlin.notifications.internal.DeletedObjectImpl
 import io.realm.kotlin.notifications.internal.PendingObjectImpl
 import io.realm.kotlin.query.RealmSingleQuery
 import io.realm.kotlin.types.BaseRealmObject
+import kotlinx.coroutines.channels.ProducerScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.filter
-import kotlinx.coroutines.flow.flatMapMerge
+import kotlinx.coroutines.flow.distinctUntilChangedBy
+import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.onStart
 import kotlin.reflect.KClass
 
 internal class SingleQuery<E : BaseRealmObject> constructor(
@@ -34,7 +36,7 @@ internal class SingleQuery<E : BaseRealmObject> constructor(
     private val classKey: ClassKey,
     private val clazz: KClass<E>,
     private val mediator: Mediator
-) : RealmSingleQuery<E>, InternalDeleteable, Thawable<Observable<RealmResultsImpl<E>, ResultsChange<E>>> {
+) : RealmSingleQuery<E>, InternalDeleteable, NotificationFlowable<RealmResultsImpl<E>, ResultsChange<E>> {
 
     override fun find(): E? {
         val link: Link = RealmInterop.realm_query_find_first(queryPointer) ?: return null
@@ -57,37 +59,36 @@ internal class SingleQuery<E : BaseRealmObject> constructor(
      * If there is an update, we ignore it, as the object flow would automatically emit the event.
      */
     override fun asFlow(): Flow<SingleQueryChange<E>> {
-        realmReference.checkClosed()
-
-        var head: E? = null
-        var headFlow: Cancellable? = null
-
+        var oldHead: E? = null
         return realmReference.owner.registerObserver(this)
-            .filter { resultsChange: ResultsChange<E> ->
-                // This filter prevents flat mapping an object flow if the object is the same.
-                val newHead: E? = resultsChange.list.firstOrNull()
-
-                val isSameObject = newHead != null && !newHead.hasSameObjectKey(head)
-                val pendingObject = resultsChange is InitialResults<E> && resultsChange.list.isEmpty()
-
-                (isSameObject || pendingObject).also {
-                    head = newHead
-                }
-            }.flatMapMerge { resultsChange ->
-                // Head was changed, cancel any active flow unless the head was deleted. In the case
-                // the head was deleted the flow would emit a [DeletedObject] and terminate.
-                if (resultsChange is UpdatedResults<*> && !resultsChange.deletions.contains(0))
-                    headFlow?.cancel()
-
-                if (resultsChange is InitialResults<E> && resultsChange.list.isEmpty()) {
-                    flowOf(PendingObjectImpl())
+            // Convert into flow of result head
+            .map { resultChange: ResultsChange<E> -> resultChange.list.firstOrNull() }
+            // Only react when head is changed
+            .distinctUntilChangedBy { head: BaseRealmObject? ->
+                head?.runIfManaged { RealmInterop.realm_object_get_key(this.objectPointer) }
+            }
+            // When head is changed issue flow that emits
+            .flatMapLatest { newHead ->
+                val oldHeadDeleted =
+                    oldHead != null && (
+                        newHead == null || RealmInterop.realm_object_resolve_in(
+                            oldHead!!.realmObjectReference!!.objectPointer,
+                            newHead.realmObjectReference!!.owner.dbPointer
+                        ) == null
+                        )
+                if (newHead == null) {
+                    if (!oldHeadDeleted) {
+                        flowOf(PendingObjectImpl())
+                    } else {
+                        flowOf(DeletedObjectImpl())
+                    }
                 } else {
-                    resultsChange.list
-                        .first()
-                        .asFlow()
-                        .also {
-                            headFlow = it as Cancellable
-                        }
+                    oldHead = newHead
+                    if (!oldHeadDeleted) {
+                        newHead.asFlow()
+                    } else {
+                        newHead.asFlow().onStart { emit(DeletedObjectImpl()) }
+                    }
                 }
             }
     }
@@ -96,8 +97,12 @@ internal class SingleQuery<E : BaseRealmObject> constructor(
      * Thaw the frozen query result, turning it back into a live, thread-confined RealmResults.
      * The results object is then used to fetch the object with index 0, which can be `null`.
      */
-    override fun thaw(liveRealm: RealmReference): RealmResultsImpl<E> =
-        thawResults(liveRealm, RealmInterop.realm_query_find_all(queryPointer), classKey, clazz, mediator)
+    override fun observable(
+        liveRealm: LiveRealm,
+        channel: ProducerScope<ResultsChange<E>>
+    ): NotificationFlow<RealmResultsImpl<E>, ResultsChange<E>> {
+        return thawResults(liveRealm.realmReference, RealmInterop.realm_query_find_all(queryPointer), classKey, clazz, mediator).observable(liveRealm, channel)
+    }
 
     override fun delete() {
         // TODO C-API doesn't implement realm_query_delete_all so just fetch the result and delete