+++ packages/test-base/src/androidAndroidTest/kotlin/io/realm/kotlin/test/shared/notifications/BacklinksNotificationsTests.kt
@@ -23,13 +23,15 @@ import io.realm.kotlin.notifications.InitialResults
 import io.realm.kotlin.notifications.ResultsChange
 import io.realm.kotlin.notifications.UpdatedResults
 import io.realm.kotlin.query.RealmResults
-import io.realm.kotlin.test.NotificationTests
+import io.realm.kotlin.test.RealmEntityNotificationTests
 import io.realm.kotlin.test.platform.PlatformUtils
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.filterNot
 import kotlinx.coroutines.flow.onCompletion
 import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.withTimeout
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
 import kotlin.test.Ignore
@@ -40,8 +42,9 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
 import kotlin.test.fail
+import kotlin.time.Duration.Companion.seconds
 
-class BacklinksNotificationsTests : NotificationTests {
+class BacklinksNotificationsTests : RealmEntityNotificationTests {
 
     lateinit var tmpDir: String
     lateinit var configuration: RealmConfiguration
@@ -238,7 +241,7 @@ class BacklinksNotificationsTests : NotificationTests {
     }
 
     @Test
-    override fun deleteObservable() {
+    override fun deleteEntity() {
         runBlocking {
             listOf(
                 realm.write { copyToRealm(Sample()) }.let { Pair(it, it.objectBacklinks) },
@@ -299,6 +302,30 @@ class BacklinksNotificationsTests : NotificationTests {
 
                 subqueryObserver.cancel()
                 sc.close()
+            }
+        }
+    }
+
+    @Test
+    override fun asFlowOnDeleteEntity() {
+        runBlocking {
+            val sample = realm.write { copyToRealm(Sample()) }
+            val mutex = Mutex(true)
+            val flow = async { sample.objectBacklinks.asFlow().collect { mutex.unlock() } }
+
+            // Await that flow is actually running
+            mutex.lock()
+            // And delete containing entity
+            realm.write { delete(findLatest(sample)!!) }
+
+            // Await that notifier has signalled the deletion so we are certain that the entity
+            // has been deleted
+            flow.await()
+
+            // Verify that a flow on the deleted entity will signal a deletion and complete gracefully
+            withTimeout(10.seconds) {
+                // First and only change should be a deletion event
+                sample.objectBacklinks.asFlow().collect { fail("Flow on deleted backlinks shouldn't emit any events") }
             }
         }
     }